+++ polardb-20170801/alibabacloud_polardb20170801/models/_add_polar_fs_quota_request.py
@@ -60,8 +60,6 @@ class AddPolarFsQuotaRequest(DaraModel):
 
         return self
 
-
-
 class AddPolarFsQuotaRequestQuotas(DaraModel):
     def __init__(
         self,